+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Fixed a bug that prevented CP users refunding orders even if they had correct permissions. ([#1259](https://github.com/craftcms/commerce/issues/1259))
 - Fixed an error that could occur when using PHP 7.0. ([#1262](https://github.com/craftcms/commerce/issues/1262))
+- Fixed a bug where selecting save and return to orders would not redirect back to the Orders index. ([#1266](https://github.com/craftcms/commerce/issues/1266))
 
 ## 3.0.5 - 2020-01-31